Reformatted JIT code
